Rename interfaces, split up code
