+++ flow/connectors/postgres/sink_pg.go
@@ -14,20 +14,6 @@ import (
 	"github.com/PeerDB-io/peer-flow/shared"
 )
 
-type QuerySinkWriter interface {
-	Close(error)
-	ExecuteQueryWithTx(context.Context, *QRepQueryExecutor, pgx.Tx, string, ...interface{}) (int, error)
-}
-
-type QuerySinkReader interface {
-	GetColumnNames() []string
-	CopyInto(context.Context, *PostgresConnector, pgx.Tx, pgx.Identifier) (int64, error)
-}
-
-type RecordStreamSink struct {
-	*model.QRecordStream
-}
-
 type PgCopyShared struct {
 	schemaLatch chan struct{}
 	schema      []string
@@ -49,86 +35,6 @@ func NewPgCopyPipe() (PgCopyReader, PgCopyWriter) {
 	schema := PgCopyShared{schemaLatch: make(chan struct{})}
 	return PgCopyReader{PipeReader: read, schema: &schema},
 		PgCopyWriter{PipeWriter: write, schema: &schema}
-}
-
-func (stream RecordStreamSink) ExecuteQueryWithTx(
-	ctx context.Context,
-	qe *QRepQueryExecutor,
-	tx pgx.Tx,
-	query string,
-	args ...interface{},
-) (int, error) {
-	defer shared.RollbackTx(tx, qe.logger)
-
-	if qe.snapshot != "" {
-		_, err := tx.Exec(ctx, "SET TRANSACTION SNAPSHOT "+QuoteLiteral(qe.snapshot))
-		if err != nil {
-			qe.logger.Error("[pg_query_executor] failed to set snapshot",
-				slog.Any("error", err), slog.String("query", query))
-			err := fmt.Errorf("[pg_query_executor] failed to set snapshot: %w", err)
-			stream.Close(err)
-			return 0, err
-		}
-	}
-
-	randomUint, err := shared.RandomUInt64()
-	if err != nil {
-		qe.logger.Error("[pg_query_executor] failed to generate random uint", slog.Any("error", err))
-		err = fmt.Errorf("[pg_query_executor] failed to generate random uint: %w", err)
-		stream.Close(err)
-		return 0, err
-	}
-
-	cursorName := fmt.Sprintf("peerdb_cursor_%d", randomUint)
-	fetchSize := shared.FetchAndChannelSize
-	cursorQuery := fmt.Sprintf("DECLARE %s CURSOR FOR %s", cursorName, query)
-	qe.logger.Info(fmt.Sprintf("[pg_query_executor] executing cursor declaration for %v with args %v", cursorQuery, args))
-	_, err = tx.Exec(ctx, cursorQuery, args...)
-	if err != nil {
-		qe.logger.Info("[pg_query_executor] failed to declare cursor",
-			slog.String("cursorQuery", cursorQuery), slog.Any("error", err))
-		err = fmt.Errorf("[pg_query_executor] failed to declare cursor: %w", err)
-		stream.Close(err)
-		return 0, err
-	}
-
-	qe.logger.Info(fmt.Sprintf("[pg_query_executor] declared cursor '%s' for query '%s'", cursorName, query))
-
-	totalRecordsFetched := 0
-	for {
-		numRows, err := qe.processFetchedRows(ctx, query, tx, cursorName, fetchSize, stream.QRecordStream)
-		if err != nil {
-			qe.logger.Error("[pg_query_executor] failed to process fetched rows", slog.Any("error", err))
-			return 0, err
-		}
-
-		qe.logger.Info(fmt.Sprintf("[pg_query_executor] fetched %d rows for query '%s'", numRows, query))
-		totalRecordsFetched += numRows
-
-		if numRows == 0 {
-			break
-		}
-	}
-
-	qe.logger.Info("Committing transaction")
-	if err := tx.Commit(ctx); err != nil {
-		qe.logger.Error("[pg_query_executor] failed to commit transaction", slog.Any("error", err))
-		err = fmt.Errorf("[pg_query_executor] failed to commit transaction: %w", err)
-		stream.Close(err)
-		return 0, err
-	}
-
-	qe.logger.Info(fmt.Sprintf("[pg_query_executor] committed transaction for query '%s', rows = %d",
-		query, totalRecordsFetched))
-	return totalRecordsFetched, nil
-}
-
-func (stream RecordStreamSink) CopyInto(ctx context.Context, _ *PostgresConnector, tx pgx.Tx, table pgx.Identifier) (int64, error) {
-	return tx.CopyFrom(ctx, table, stream.GetColumnNames(), model.NewQRecordCopyFromSource(stream.QRecordStream))
-}
-
-func (stream RecordStreamSink) GetColumnNames() []string {
-	return stream.Schema().GetColumnNames()
 }
 
 func (p PgCopyWriter) SetSchema(schema []string) {
